Feat(pkg): Add rollback command and functionality
Introduce 'zoi rollback <package>' command to restore a package to its previously installed version.

A backup of the currently installed package is automatically created when installing a different version of an already installed package. This backup is then used by the rollback command.

The rollback functionality can be globally enabled/disabled via 'rollback_enabled' in '~/.zoi/config.yaml' and can be overridden for specific packages using the 'rollback' field in their manifests.