Config(registry): Use build-time configurable default registry
Introduces the ZOI_DEFAULT_REGISTRY environment variable, which can be set during the build process to define the application's default registry URL.
This value is prioritized from the system environment over the .env.local file. If neither is set, it defaults to a hardcoded GitLab URL.
All hardcoded default registry URLs in the application have been replaced with a call to the new pkg::config::get_default_registry() function, which retrieves this configured value.